+++ event_rule_updated.go
@@ -5,13 +5,13 @@ func init() {
 }
 
 type RuleUpdated struct {
-	RuleID                string  `json:"rule_id" analytics:"rul"`
-	UpdatedBy             string  `json:"updated_by" analytics:"usr"`
-	UsesSelectableOptions bool    `json:"uses_selectable_options"`
-	UsesDynamicOptions    bool    `json:"uses_dynamic_options"`
-	Provider              string  `json:"provider"`
-	MaxDurationSeconds    float64 `json:"max_duration_seconds"`
-	RequiresApproval      bool    `json:"requires_approval"`
+	RuleID                string `json:"rule_id" analytics:"rul"`
+	UpdatedBy             string `json:"updated_by" analytics:"usr"`
+	UsesSelectableOptions bool   `json:"uses_selectable_options"`
+	UsesDynamicOptions    bool   `json:"uses_dynamic_options"`
+	Provider              string `json:"provider"`
+	MaxDurationSeconds    int    `json:"max_duration_seconds"`
+	RequiresApproval      bool   `json:"requires_approval"`
 }
 
 func (r *RuleUpdated) userID() string { return r.UpdatedBy }
